+++ src/main/java/com/github/buerxixi/easydbf/model/DBFDateField.java
@@ -12,6 +12,6 @@ import com.github.buerxixi.easydbf.pojo.DBFFieldType;
 public class DBFDateField extends DBFField {
 
     public DBFDateField(String name) {
-        super(DBFConstant.DEFAULT_CHARSET, name, DBFFieldType.DATE,-1, (byte) 8, (byte) 0);
+        super(DBFConstant.DEFAULT_CHARSET, name, DBFFieldType.DATE,-1, 8, 0);
     }
 }
